Added more levels (#94)
* Cave level

* Quick update

* Template update

* Fixed amount of levels

* Lobby help text visible

* Reworked volcano to lava and added animation layer

* Moved torch out of stone background in lava.json



* Fixed wrongly displacing weapons

* Big town level

* City done

* Cave fix

* Engine fix

* More engine update

* Highscore controller fix

* added spawn direction

* More cave fixerino

* Fixed player falling though everything

* smollfix

* Delete volcano.json

* Start of level2.json

* bobbo

* fixed some levels

* More cave fix

* Added layers and direction converter

* Updated engine

* Added pirate level

+++ lib/systems/highscore_system.cpp
@@ -25,7 +25,7 @@ void HighscoreSystem::update(double) {
     }
     if(initialized) {
         auto& input = BrickInput<PlayerInput>::getInstance();
-        int x = input.checkInput(1, PlayerInput::X_AXIS);
+        double x = input.checkInput(1, PlayerInput::X_AXIS);
         if(x > 0) {
             entityManager->removeEntitiesWithTag("HighscoreScene_player");
             selector++;
